# src/api-spec/openapi.yaml
openapi: 3.0.3
info:
  title: PySCF_frontlication API
  description: |
    Backend API for managing quantum chemistry calculations using PySCF.
    
    ## Discriminated Union Pattern for Calculation Requests
    
    This API uses **discriminated unions** (also known as tagged unions) to provide type-safe calculation request schemas.
    Each calculation method (DFT, HF, MP2, CCSD, CCSD(T), TDDFT, CASCI, CASSCF) has its own dedicated schema with 
    method-specific parameters.
    
    ### Design Principles
    
    1. **Type Safety**: The `calculation_method` field acts as a discriminator that determines which schema applies
    2. **Single Source of Truth**: Each method's parameters are defined once in the schema and enforced by backend validation
    3. **Method-Specific Parameters**: Parameters like `exchange_correlation` (DFT/TDDFT only), `ncas` (CASCI/CASSCF only), 
       or `frozen_core` (CCSD/CCSD(T) only) are only present in their applicable schemas
    4. **Explicit Constraints**: Disabled parameters (e.g., `optimize_geometry` for TDDFT) use `enum: [false]` to enforce restrictions
    
    ### Schema Structure
    
    ```yaml
    QuantumCalculationRequest:
      oneOf:
        - $ref: '#/components/schemas/DFTCalculationRequest'
        - $ref: '#/components/schemas/HFCalculationRequest'
        - $ref: '#/components/schemas/MP2CalculationRequest'
        # ... (all 8 method-specific schemas)
      discriminator:
        propertyName: calculation_method
    ```
...
servers:
...
paths:
...
  /api/quantum/calculate:
    post:
      summary: Start quantum chemistry calculation
      description: Start a calculation in the background
      operationId: startCalculation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuantumCalculationRequest'
      responses:
        '202':
          description: Calculation queued successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StartCalculationResponse'
        '400':
          description: Invalid request parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Failed to queue calculation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/quantum/calculations:
    get:
      summary: List calculations with optional filtering
      description: Get a list of calculation directories with optional filtering by name, status, method, basis set, and date range. Filtering is performed at the file system level for optimal performance.
      operationId: listCalculations
      parameters:
        - name: name_query
          in: query
          required: false
          schema:
            type: string
          description: Partial match search in calculation name (case-insensitive)
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [pending, running, completed, error, waiting, pausing, paused]
          description: Filter by calculation status
        - name: calculation_method
          in: query
          required: false
          schema:
            type: string
            enum: [DFT, HF, MP2, CCSD, TDDFT, CASCI, CASSCF]
          description: Filter by calculation method
        - name: basis_function
          in: query
          required: false
          schema:
            type: string
          description: Filter by basis set (case-insensitive)
        - name: date_from
          in: query
          required: false
          schema:
            type: string
            format: date
          description: Start date for date range filtering (ISO format YYYY-MM-DD)
        - name: date_to
          in: query
          required: false
          schema:
            type: string
            format: date
          description: End date for date range filtering (ISO format YYYY-MM-DD)
      responses:
        '200':
          description: List of calculations matching filter criteria
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CalculationListResponse'
        '500':
          description: Failed to list calculations
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/quantum/calculations/{calculationId}:
    get:
      summary: Get calculation details
      description: Get detailed information about a specific calculation
      operationId: getCalculationDetails
      parameters:
        - name: calculationId
          in: path
          required: true
          description: Unique calculation ID
          schema:
            type: string
      responses:
        '200':
          description: Calculation details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CalculationDetailsResponse'
        '404':
          description: Calculation not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Failed to get calculation details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    patch:
      summary: Update calculation metadata
      description: Update calculation name and other metadata
      operationId: patchCalculation
      parameters:
        - name: calculationId
          in: path
          required: true
          description: Unique calculation ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CalculationUpdateRequest'
      responses:
        '200':
          description: Calculation updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CalculationUpdateResponse'
        '404':
          description: Calculation not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Name conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Failed to update calculation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      summary: Delete calculation
      description: Delete a calculation and all its associated files
      operationId: deleteCalculation
      parameters:
        - name: calculationId
          in: path
          required: true
          description: Unique calculation ID
          schema:
            type: string
      responses:
        '200':
          description: Calculation deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CalculationDeleteResponse'
        '404':
          description: Calculation not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Failed to delete calculation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/quantum/calculations/{calculationId}/pause:
    post:
      summary: Pause a running calculation
      description: Safely pause a running calculation after the current SCF iteration completes. The calculation can be resumed later from the checkpoint.
      operationId: pauseCalculation
      parameters:
        - name: calculationId
          in: path
          required: true
          description: Unique calculation ID
          schema:
            type: string
      responses:
        '202':
          description: Pause request accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PauseCalculationResponse'
        '400':
          description: Calculation is not in a pausable state (not running)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Calculation not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Failed to pause calculation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/quantum/calculations/{calculationId}/resume:
    post:
      summary: Resume a paused calculation
      description: Resume a paused calculation from its checkpoint file. The calculation will continue from where it was paused.
      operationId: resumeCalculation
      parameters:
        - name: calculationId
          in: path
          required: true
          description: Unique calculation ID
          schema:
            type: string
      responses:
        '202':
          description: Resume request accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResumeCalculationResponse'
        '400':
          description: Calculation cannot be resumed (not paused, no checkpoint available, etc.)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Calculation not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Failed to resume calculation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/quantum/calculations/{calculationId}/orbitals:
    get:
      summary: Get molecular orbital information
      description: Get list of all molecular orbitals with energies and indices
      operationId: getOrbitals
      parameters:
        - name: calculationId
          in: path
          required: true
          description: Unique calculation ID
          schema:
            type: string
      responses:
        '200':
          description: Molecular orbital information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrbitalsResponse'
        '404':
          description: Calculation not found or no orbital data available
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Failed to get orbital information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/quantum/calculations/{calculationId}/orbitals/{orbitalIndex}/cube:
    get:
      summary: Generate and get molecular orbital CUBE file
      description: Generate CUBE file for specific molecular orbital visualization
      operationId: getOrbitalCube
      parameters:
        - name: calculationId
          in: path
          required: true
          description: Unique calculation ID
          schema:
            type: string
        - name: orbitalIndex
          in: path
          required: true
          description: Molecular orbital index
          schema:
            type: integer
            minimum: 0
        - name: gridSize
          in: query
          required: false
          description: Grid size for CUBE file generation (default 80)
          schema:
            type: integer
            minimum: 40
            maximum: 120
            default: 80
        - name: isovaluePos
          in: query
          required: false
          description: Positive isovalue for orbital visualization (default 0.02)
          schema:
            type: number
            minimum: 0.001
            maximum: 0.1
            default: 0.02
        - name: isovalueNeg
          in: query
          required: false
          description: Negative isovalue for orbital visualization (default -0.02)
          schema:
            type: number
            minimum: -0.1
            maximum: -0.001
            default: -0.02
      responses:
        '200':
          description: Orbital CUBE file data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrbitalCubeResponse'
        '404':
          description: Calculation or orbital not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Failed to generate orbital CUBE file
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/quantum/calculations/{calculationId}/orbitals/cube-files:
    get:
      summary: List CUBE files for a calculation
      description: List all saved CUBE files for molecular orbitals in a calculation
      operationId: listCubeFiles
      parameters:
        - name: calculationId
          in: path
          required: true
          description: Unique calculation ID
          schema:
            type: string
      responses:
        '200':
          description: CUBE files list retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CubeFilesListResponse'
        '404':
          description: Calculation not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Failed to retrieve CUBE files
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      summary: Delete CUBE files for a calculation
      description: Delete CUBE files for specific orbital or all orbitals in a calculation
      operationId: deleteCubeFiles
      parameters:
        - name: calculationId
          in: path
          required: true
          description: Unique calculation ID
          schema:
            type: string
        - name: orbitalIndex
          in: query
          required: false
          description: Specific orbital index to delete (if not provided, deletes all)
          schema:
            type: integer
      responses:
        '200':
          description: CUBE files deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CubeFilesDeleteResponse'
        '404':
          description: Calculation not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Failed to delete CUBE files
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/quantum/supported-parameters:
    get:
      summary: Get supported quantum chemistry parameters
      description: Get lists of supported calculation methods, basis functions, exchange-correlation functionals, and solvents
      operationId: getSupportedParameters
      responses:
        '200':
          description: Supported parameters retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SupportedParametersResponse'
        '500':
          description: Failed to retrieve supported parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    QuantumCalculationRequest:
      oneOf:
        - $ref: '#/components/schemas/DFTCalculationRequest'
        - $ref: '#/components/schemas/HFCalculationRequest'
        - $ref: '#/components/schemas/MP2CalculationRequest'
        - $ref: '#/components/schemas/CCSDCalculationRequest'
        - $ref: '#/components/schemas/CCSD_TCalculationRequest'
        - $ref: '#/components/schemas/TDDFTCalculationRequest'
        - $ref: '#/components/schemas/CASCICalculationRequest'
        - $ref: '#/components/schemas/CASSCFCalculationRequest'
      discriminator:
        propertyName: calculation_method
        mapping:
          DFT: '#/components/schemas/DFTCalculationRequest'
          HF: '#/components/schemas/HFCalculationRequest'
          MP2: '#/components/schemas/MP2CalculationRequest'
          CCSD: '#/components/schemas/CCSDCalculationRequest'
          CCSD_T: '#/components/schemas/CCSD_TCalculationRequest'
          TDDFT: '#/components/schemas/TDDFTCalculationRequest'
          CASCI: '#/components/schemas/CASCICalculationRequest'
          CASSCF: '#/components/schemas/CASSCFCalculationRequest'
...
    StartCalculationResponse:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          required: [calculation]
          properties:
            calculation:
              $ref: '#/components/schemas/CalculationInstance'
...